+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Change approach for injecting values into the Rmd file, to avoid the need to duplicate escapes
 * Allow the user to set the column to colour by in the PCA
+* Include 'extra' columns in output from runPTMTest
 
 # einprot 0.6.10